requirements.txt
Beaker==1.6.4
Brlapi==0.5.6
Electrum==1.7.2
GnuPGInterface==0.3.2
Jinja2==2.6
Mako==0.5.0
MarkupSafe==0.15
PAM==0.4.2
PIL==1.1.7
Pygments==1.4
Sphinx==1.1.3
Twisted==11.1.0
Twisted-Conch==11.1.0
Twisted-Core==11.1.0
Twisted-Lore==11.1.0
Twisted-Mail==11.1.0
Twisted-Names==11.1.0
Twisted-News==11.1.0
Twisted-Runner==11.1.0
Twisted-Web==11.1.0
Twisted-Words==11.1.0
adium-theme-ubuntu==0.3.2
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
bottle==0.10.11
bsddb3==5.2.0
chardet==2.0.1
command-not-found==0.2.44
configglue==1.0
configobj==4.7.2
debtagshw==0.1
decorator==3.3.2
defer==1.0.2
dirspec==3.0.0
dirtools==0.2.0
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
docutils==0.8.1
duplicity==0.6.18
ecdsa==0.8
epydoc==3.0.1
globster==0.1.0
httplib2==0.7.2
ipython==0.12.1
joblib==0.5.4.dev
jockey==0.9.7
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
lockfile==0.8
louis==2.3.0
matplotlib==1.1.1rc
nose==1.1.2
numpy==1.6.1
nvidia-common==0.0.0
oauth==1.0.1
onboard==0.97.1
oneconf==0.2.8.1
pexpect==2.3
piston-mini-client==0.7.2
protobuf==2.4.1
psycopg2==2.4.5
pyOpenSSL==0.12
pyasn1==0.0.11a
pycrypto==2.4.1
pycups==1.9.61
pycurl==7.19.0
pyinotify==0.9.2
pyparsing==1.5.2
pyserial==2.5
pysmbc==1.0.13
python-apt==0.8.3ubuntu7.1
python-daemon==1.5.5
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-virtkey==0.60.0
pythonbrew==1.3
pytz==2011k
pyxdg==0.19
reportlab==2.5
rhythmbox-ubuntuone==3.0.0
scikit-learn==0.10
scipy==0.9.0
sessioninstaller==0.0.0
simplegeneric==0.7
simplejson==2.3.2
slowaes==0.1a1
software-center-aptd-plugins==0.0.0
system-service==0.1.6
traits==4.0.0
ubuntuone-couch==0.3.0
ubuntuone-installer==3.0.2
ubuntuone-storage-protocol==3.0.2
ufw==0.31.1-1
unattended-upgrades==0.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
usb-creator==0.2.23
virtualenv==1.10.1
wadllib==1.3.0
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==2.5.3
xkit==0.0.0
xlwt==0.7.2
zope.interface==3.6.1
